move common code to a base class for ELM327 and PiOBD2Hat to ease maintenance
