+++ dongles/ELM327.py
@@ -1,107 +1,12 @@
-from dongle import *
-from serial import Serial
-import logging
-import math
-import pexpect
-from pexpect import fdpexpect
+from at-base-dongle import *
 
-class ELM327:
+class ELM327(ATBASE):
     def __init__(self, dongle, watchdog = None):
-        print("Init Dongle")
-        self.serial = Serial(dongle['port'], baudrate=dongle['speed'], timeout=5)
-        self.exp = fdpexpect.fdspawn(self.serial.fd)
-        self.initDongle()
-
-        self.config = dongle
-        self.watchdog = watchdog
-
-    def sendAtCmd(self, cmd, expect='OK'):
-        cmd = bytes(cmd, 'utf-8')
-        expect = bytes(expect, 'utf-8')
-        try:
-            while self.serial.in_waiting:   # Clear the input buffer
-                print("Stray data in buffer: " + \
-                        str(self.serial.read(self.serial.in_waiting)))
-                sleep(0.2)
-            self.exp.send(cmd + b'\r\n')
-            self.exp.expect('>')
-            ret = self.exp.before.strip(b'\r\n')
-            if expect not in ret:
-                raise Exception('Expected %s, got %s' % (expect,ret))
-
-        except pexpect.exceptions.TIMEOUT:
-            ret = b'TIMEOUT'
-
-        return ret.split(b"\r\n")[-1]
-
-        expect = bytes(expect, 'utf-8')
-        cmd = bytes(cmd, 'utf-8')
-
-    def sendCommand(self, cmd):
-        """
-        @cmd: should be hex-encoded
-        """
-
-        cmd = cmd.hex()
-        try:
-            while self.serial.in_waiting:   # Clear the input buffer
-                print("Stray data in buffer: " + \
-                        str(self.serial.read(self.serial.in_waiting)))
-                sleep(0.2)
-            self.exp.send(cmd + '\r\n')
-            self.exp.expect('>')
-            ret = self.exp.before.strip(b'\r\n')
-        except pexpect.exceptions.TIMEOUT:
-            ret = b'TIMEOUT'
-
-        if ret in [b'NO DATA', b'DATA ERROR', b'ACT ALERT']:
-            raise NoData(ret)
-        elif ret in [b'BUFFER FULL', B'BUS BUSY', b'BUS ERROR', b'CAN ERROR',
+        ATBASE.__init__(self, dongle, watchdog)
+        self.ret_NoData = (b'NO DATA', b'DATA ERROR', b'ACT ALERT')
+        self.ret_CanError = (b'BUFFER FULL', B'BUS BUSY', b'BUS ERROR', b'CAN ERROR',
                 b'ERR', b'FB ERROR', b'LP ALERT', b'LV RESET', b'STOPPED',
-                b'UNABLE TO CONNECT']:
-            raise CanError("Failed Command {}\n{}".format(cmd,ret))
-
-        try:
-            data = {}
-            raw = ret.split(b'\r\n')
-            lines = None
-
-            for line in raw:
-                if len(line) != 19:
-                    raise ValueError
-
-                identifier = int(line[0:3],16)
-                frame_type = int(line[3:4],16)
-
-                if frame_type == 0:     # Single frame
-                    idx = 0
-                    lines = 1
-                elif frame_type == 1:   # First frame
-                    lines = math.ceil((int.from_bytes(bytes.fromhex(str(b'0' + line[4:7], 'ascii')),
-                        byteorder='big', signed=False) + 1) / 7)
-                    idx = 0
-                    if len(raw) != lines:
-                        raise ValueError
-                elif frame_type == 2:   # Consecutive frame
-                    idx = int(line[4:5],16)
-                else:                   # Unexpected frame
-                    raise ValueError
-
-                if not identifier in data:
-                    data[identifier] = [None] * lines
-
-                data[identifier][idx] = bytes.fromhex(str(line[5:],'ascii'))
-
-            # Check if all entries are filled
-            for key, val in data.items():
-                for i in val:
-                    if i == None:
-                        raise ValueError
-
-        except ValueError:
-            raise CanError("Failed Command {}\n{}".format(cmd,ret))
-
-        return data
+                b'UNABLE TO CONNECT')
 
     def initDongle(self):
         cmds = [['ATZ','OK'],
@@ -115,12 +20,13 @@ class ELM327:
         for c,r in cmds:
             self.sendAtCmd(c, r)
 
-    def setAllowLongMessages(self, value):
-        self.sendAtCmd('ATAL%i' % value)
-
     def setProtocol(self, prot):
         if prot == 'CAN_11_500':
             self.sendAtCmd('ATSP6','OK')
+            self.is_extended = False
+        elif prot == 'CAN_29_500':
+            self.sendAtCmd('ATSP7','OK')
+            self.is_extended = True
         else:
             raise Exception('Unsupported protocol %s' % prot)
 
@@ -128,25 +34,31 @@ class ELM327:
         if isinstance(can_id, bytes):
             can_id = str(can_id)
         elif isinstance(can_id, int):
-            can_id = format(can_id, 'X')
+            can_id = format(can_id, '08X' if self.is_extended else '03X')
 
-        self.sendAtCmd('ATSH' + can_id)
+        if self.current_canid != can_id:
+            self.sendAtCmd('ATSH' + can_id)
+            self.current_canid == can_id
 
     def setCANRxMask(self, mask):
         if isinstance(mask, bytes):
             mask = str(mask)
         elif isinstance(mask, int):
-            mask = format(mask, 'X')
+            mask = format(mask, '08X' if self.is_extended else '03X')
 
-        self.sendAtCmd('ATCM' + mask)
+        if self.current_canmask != mask:
+            self.sendAtCmd('ATCM' + mask)
+            self.current_canmask == mask
 
     def setCANRxFilter(self, addr):
         if isinstance(addr, bytes):
             addr = str(addr)
         elif isinstance(addr, int):
-            addr = format(addr, 'X')
+            addr = format(addr, '08X' if self.is_extended else '03X')
 
-        self.sendAtCmd('ATCF' + addr)
+        if self.current_canfilter != addr:
+            self.sendAtCmd('ATCF' + addr)
+            self.current_canfilter == addr
 
     def getObdVoltage(self):
         if self.watchdog:
@@ -154,12 +66,6 @@ class ELM327:
         else:
             ret = self.sendAtCmd('ATRV')
             return round(float(ret[:-1]), 2)
-
-    def isCarAvailable(self):
-        if self.watchdog:
-            return self.watchdog.getShutdownFlag() == 0
-        else:
-            return self.getObdVoltage() > 13.0
 
     def calibrateObdVoltage(self, realVoltage):
         if self.watchdog:
